+++ matrix-client/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Replaces MatrixIO with a new mtl style API using ExceptT MatrixError and ReaderT ClientSession.
 - Adds loginToken.
 - Adds loginTokenWithManager and createSessionWithManager to support custom http Manager.
+- Replaces IdentitySession with ClientSession.
 
 ## 0.1.4.0